% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-missing.R
\name{code_missing}
\alias{code_missing}
\title{Replace Values not coded with NA}
\usage{
code_missing(x, value = -1)
}
\arguments{
\item{x}{Vector of Data}

\item{value}{Used when replacing observation with a missing entry.
Default is \code{-1}.}
}
\value{
A \code{vector} with missing values appropriately coded.
}
\description{
Modifies the data by searching for what the researcher specified as
the missing value and substitutes \emph{R}'s missing data type \code{NA}.
}
\examples{
# Data
x = c(1, 2, -1, 4)

# Replace -1 with NA
y = code_missing(x)

# Replace 4 with NA
y = code_missing(x, value = 4)
}
\author{
James Joseph Balamuta
}
