% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotchicken-package.R
\docType{package}
\name{hotchicken-package}
\alias{hotchicken}
\alias{hotchicken-package}
\title{hotchicken: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: James Joseph Balamuta \email{balamut2@illinois.edu} (\href{https://orcid.org/0000-0003-2826-8458}{ORCID}) [copyright holder]

}
\keyword{internal}
